[v2] move withPrefix to gatsby package
