+++ packages/gatsby-plugin-feed/src/gatsby-ssr.js
@@ -1,5 +1,5 @@
 import React from "react"
-import { withPrefix } from "gatsby-link"
+import { withPrefix } from "gatsby"
 import { defaultOptions } from "./internals"
 
 exports.onRenderBody = ({ setHeadComponents }, pluginOptions) => {
