+++ packages/gatsby/README.md
@@ -87,21 +87,21 @@ hosted on GitHub Pages or other places where your site isn't at the root of the
 domain.
 
 This can create problems during development as pathnames won't be prefixed. To
-handle both, gatsby-link exports a helper function `withPrefix` that prepends
+handle both, gatsby exports a helper function `withPrefix` that prepends
 the prefix during production but doesn't in development.
-
-This is only for pathnames you're constructing manually. The `<Link>` component
-handles prefixing automatically.
 
 ```jsx
 import { withPrefix } from "gatsby"
+import Helmet from "react-helmet"
 
-const IndexLayout = ({ children, location }) => {
+const favicon = withPrefix("/favicon.png")
+
+const Layout = ({ children }) => {
   const isHomepage = location.pathname === withPrefix("/")
 
   return (
     <div>
-      <h1>Welcome {isHomepage ? "home" : "aboard"}!</h1>
+      <Helmet link={[{ rel: "shortcut icon", href: favicon }]} />
       {children}
     </div>
   )
