+++ packages/gatsby/src/cache-dir/prefix.js
@@ -3,7 +3,11 @@
  * isn't found.
  */
 
-export default (str, prefix = ``) => {
+export function stripPrefix(str, prefix = ``) {
   if (str.substr(0, prefix.length) === prefix) return str.slice(prefix.length)
   return str
+}
+
+export function withPrefix(path) {
+  return `${__PATH_PREFIX__}/${path}`.replace(/\/+/g, `/`)
 }
